Move type traits into a separate file
